Struct chess::board::Board

pub struct Board {
    color: Color,
    pub board: Vec<Vec<Option<Piece>>>,
}

Represents the board state. color represents the turn i.e. white indicates it is white's turn. The board is represented by an 8x8 matrix of Option<Piece>.

Fields

color

board

Methods

impl Board

fn symmetrical_pieces(black_i: i32, j: i32, board: &mut Vec<Vec<Option<Piece>>>, piece_type: PieceType)

Helper function to put the four symmetrical pieces on the board.

fn new() -> Board

Initialize the board with starting positions.

fn white_turn(&self) -> bool

Helper function that checks if it is white's turn.

fn get_piece(&self, cell: &Cell) -> &Option<Piece>

Get the piece associated with a given cell index.

fn inbounds(&self, cell: &Cell) -> bool

Helper function to check if a proposed cell is in the bounds of the board.

fn is_empty(&self, cell: &Cell) -> bool

Helper function to check if a cell has a piece.

fn is_enemy(&self, from: &Cell, to: &Cell) -> bool

Helper function to check if a cell holds an enemy piece.

fn is_friendly_board(&self, cell: &Cell) -> bool

Helper function to check if a cell holds a friendly piece relative to board.

fn moves_until_collision(&self, dirs: Vec<(i32, i32)>, cell: Cell) -> Vec<Cell>

Returns the potential moves in the given directions, stopping upon a collision. Used for Bishop, Rook, and Queen.

fn pawn_moves(&self, cell: Cell) -> Vec<Cell>

Helper function to implement pawn logic.

fn basic_moves(&self, dirs: Vec<(i32, i32)>, cell: Cell) -> Vec<Cell>

Helper function to implement logic for pieces with set directions. Pieces include the Knight and the King.

fn potential_moves(&self, cell: &Cell) -> Vec<Cell>

Calculate the potential moves for a given cell index.

fn self_check(&self, from: Cell, to: Cell) -> bool

Helper function to check if a move would place the current player's king in danger. Clones the board and makes the move, before running check() on the clone.

fn switch_color(&mut self)

Helper function to swap the current player.

fn friendly_pieces(&self) -> Vec<Cell>

Helper function to return cells of all friendly pieces.

fn enemy_pieces(&self) -> Vec<Cell>

Helper function to return cells of all enemy pieces.

fn check(&self) -> bool

Helper function that checks for check, i.e. if the enemy king is in danger. Iterates through all friendly pieces and checks if the enemy king is in any of their potential moves.

fn checkmate(&self) -> bool

Helper function that checks for checkmate. Iterates through all enemy pieces and checks if any of their potential moves can bring them out of check.

fn move_piece(&mut self, from: Cell, to: Cell) -> bool

Helper function that moves a piece from a cell to the target cell

fn contains_enemy_king(&self, cells: Vec<Cell>) -> bool

Helper function that checks if a list of potential moves contains the enemy king.