Module chess::chess_server

A module for the server logic of Chess. Handles all WebSocket clients using a relay MPSC channel. Sends information about game logic through Actions.

Reexports

use std::prelude::v1::*;

use rustc_serialize::json;

use std::str;

use std::thread;

use std::sync::{Arc, Mutex, mpsc};

use std::collections::HashMap;

use websocket::{Receiver, Sender, Server, Message, WebSocketStream};

use websocket::sender;

use websocket::receiver;

use websocket::message::Type;

use super::board::{Board, Cell};

Structs

Payload

Enums

Action
Represents a single, atomic action taken by a client.

Constants

WS_ADDR

Functions

client_thread
Each client thread waits for input (or disconnects) from its respective clients and relays the appropriate messages via the relay MPSC channel.

listen
Create the relay MPSC (multi-producer/single-consumer) channel, spawn the relay thread, then listen for WebSocket clients and spawn their threads.

relay_thread
The relay thread handles all Actions received on its MPSC channel. If it receives an 'Action::Select', it will calculate the potential_moves that can be made by that cell, and then relay an 'Action::Moves' to the sender only. If it receives an 'Action::Move', it will adjust the board state, check for check and checkmate, switch turns, and then send an 'Action::Board' with all the state to all clients. If it receives any other 'Action', it will relay the 'Action' verbatim to all clients.

start
Spawn a WebSocket listener thread.