Function chess::chess_server::client_thread

fn client_thread(mutex_board: Arc<Mutex<Board>>, black_ip: Arc<Mutex<String>>, white_ip: Arc<Mutex<String>>, ip: String, mpsc_sender: Sender<String>, client_receiver: Receiver<WebSocketStream>)

Each client thread waits for input (or disconnects) from its respective clients and relays the appropriate messages via the relay MPSC channel.

The messages received-from and sent-to the client should be JSON objects with the same form as rustc_serialize's serialization of the Action type.

If the client connects, a Action::Connect will be relayed with their IP address.

If the client disconnects, a Action::Disconnect will be relayed with their IP address.

If the client sends a 'Action::Msg', the message will be relayed and the user will be adjusted to Black, White, or Spectator accordingly.

If the client sends a 'Action::Select', if it is not their turn then this message will be discarded. Otherwise, the 'Action::Select' will be relayed with the selected cell and the client's IP address.

If the client sends a 'Action::Move', if it is not their turn then this message will be discarded. Otherwise, the 'Action::Move' will be relayed with the from cell and to cell.