Enum chess::CastleRights

pub enum CastleRights {
    NoRights,
    KingSide,
    QueenSide,
    Both,
}

What castle rights does a particular player have?

Variants
NoRightsKingSideQueenSideBoth
Methods

impl CastleRights

fn has_kingside(&self) -> bool

Can I castle kingside?

fn has_queenside(&self) -> bool

Can I castle queenside?

fn kingside_squares(&self, color: Color) -> BitBoard

What squares need to be empty to castle kingside?

fn queenside_squares(&self, color: Color) -> BitBoard

What squares need to be empty to castle queenside?

fn remove(&self, remove: CastleRights) -> CastleRights

Remove castle rights, and return a new CastleRights.

fn add(&self, add: CastleRights) -> CastleRights

Add some castle rights, and return a new CastleRights.

fn to_index(&self) -> usize

Convert CastleRights to usize for table lookups

fn from_index(i: usize) -> CastleRights

Convert usize to CastleRights. Panic if invalid number.

fn unmoved_rooks(&self, color: Color) -> BitBoard

Which rooks can we "guarantee" we haven't moved yet?

fn rook_square_to_castle_rights(square: Square) -> CastleRights

Given a square of a rook, which side is it on? Note: It is invalid to pass in a non-rook square. The code may panic.