Function chess::pos_to_string

pub fn pos_to_string(pos: (usize, usize)) -> Result<String, i32>

Turns a position on the board from a tuple, like (3, 5), to proper chess notation, like D6.

Returns a Result containing the string, or an error if the given tuple was out of bounds.

Eksamples

assert_eq!(pos_to_string((3, 5)), Ok("D6".to_string()));
assert_eq!(pos_to_string((0, 0)), Ok("A1".to_string()));
assert_eq!(pos_to_string((7, 7)), Ok("H8".to_string()));

// Returns Err(1) when the values are out of bounds.
assert_eq!(pos_to_string((8, 8)), Err(1));