chess

Module make_move

Functions

apply_move_unchecked

Applies a move directly to the given position without legality checks.

is_checked

Returns true if the given color’s king is in check.

is_checkmated

Returns true if the given color is checkmated.

is_stalemated

Returns true if the given color is stalemated.

legal_moves

Generates all legal moves for the given color.

make_move

Attempts to make a move in the given game.

update_castling_rights 🔒

Updates castling rights in the given position after a move.