chess::make_move

Function legal_moves

pub fn legal_moves(color: Color, position: &Position) -> Vec<Move>

Generates all legal moves for the given color.

Pseudo-legal moves are generated with valid_moves, then filtered to exclude moves that leave the king in check.

Arguments

color - The side to generate moves for.

position - The board state.

Returns

A vector of all legal moves available to color.