Struct chess::CacheTable

pub struct CacheTable<T: Copy + Clone + PartialEq + PartialOrd> { /* fields omitted */ }

Store a cache of entries, each with an associated hash.

Methods

impl<T: Copy + Clone + PartialEq + PartialOrd> CacheTable<T>

fn new(size: usize, default: T) -> CacheTable<T>

Create a new CacheTable with each associated entry initialized with a hash of '0' Note: You must pass in a size where only 1 bit is set. (AKA: 2, 4, 8, 16, 1024, 65536, etc.) Panics when size is invalid.

fn get(&self, hash: u64) -> Option<T>

Get a particular entry with the hash specified

fn add(&mut self, hash: u64, entry: T)

Add (or overwrite) an entry with the associated hash