Struct chess::Square

pub struct Square(_);

Represent a square on the chess board

Methods

impl Square

unsafe fn new(sq: u8) -> Square

Create a new square, given an index. Note: It is invalid, but allowed, to pass in a number >= 64. Doing so will crash stuff.

fn make_square(rank: Rank, file: File) -> Square

Make a square given a rank and a file

fn get_rank(&self) -> Rank

Return the rank given this square.

fn get_file(&self) -> File

Return the file given this square.

fn up(&self) -> Option<Square>

If there is a square above me, return that. Otherwise, None.

fn down(&self) -> Option<Square>

If there is a square below me, return that. Otherwise, None.

fn left(&self) -> Option<Square>

If there is a square to the left of me, return that. Otherwise, None.

fn right(&self) -> Option<Square>

If there is a square to the right of me, return that. Otherwise, None.

fn forward(&self, color: Color) -> Option<Square>

If there is a square "forward", given my Color, go in that direction. Otherwise, None.

fn backward(&self, color: Color) -> Option<Square>

If there is a square "backward" given my Color, go in that direction. Otherwise, None.

fn uup(&self) -> Square

If there is a square above me, return that. If not, wrap around to the other side.

fn udown(&self) -> Square

If there is a square below me, return that. If not, wrap around to the other side.

fn uleft(&self) -> Square

If there is a square to the left of me, return that. If not, wrap around to the other side.

fn uright(&self) -> Square

If there is a square to the right of me, return that. If not, wrap around to the other side.

fn uforward(&self, color: Color) -> Square

If there is a square "forward", given my color, return that. If not, wrap around to the other side.

fn ubackward(&self, color: Color) -> Square

If there is a square "backward", given my color, return that. If not, wrap around to the other side.

fn to_int(&self) -> u8

Convert this square to an integer.

fn to_index(&self) -> usize

Convert this Square to a usize for table lookup purposes

fn from_string(s: String) -> Option<Square>

Convert a UCI String to a square. If invalid, return None