Crate chess

Reexports

extern crate std as std;

extern crate hyper;

extern crate iron;

extern crate staticfile;

extern crate mount;

extern crate rustc_serialize;

extern crate websocket;

use std::prelude::v1::*;

Modules

board
A module for the game logic of Chess. Stores the Board object and contains methods for finding potential moves, check/checkmate, etc. It utilizes Cell objects in order to index into the board, which is a Vec x Vec of Piece options. The Piece object contains an enum of PieceType, which is used to determine the game behavior for each piece.

chess_server
A module for the server logic of Chess. Handles all WebSocket clients using a relay MPSC channel. Sends information about game logic through Actions.

page

Functions

main