Crate chess

A basic chess backend.

Eksamples

let mut game = Game::new();
// Move white pawn from E2 to E4
game.move_piece((4, 1), (4, 3));

// Move black pawn from D7 to D5
game.move_piece((3, 6), (3, 4));

// Move white pawn from E4 to D5
let captured = game.move_piece((4, 3), (3, 4));
// Check that the black pawn was captured
match captured {
    Some(piece) => {
        assert_eq!(piece.kind, Kind::Pawn);
        assert_eq!(piece.color, Color::Black);
    },
    None => panic!("Expected a black pawn"),
}

Structs

Game
The game struct.

Piece
The chess piece struct.

Enums

Color
The different colors of chess pieces.

Kind
The different kinds of chess pieces.

Victory
The different types of victories.

Statics

BLACK
An array of all the black chess pieces.

WHITE
An array of all the white chess pieces.

Functions

pos_to_string
Turns a position on the board from a tuple, like (3, 5), to proper chess notation, like D6.

string_to_pos
Turns a position on the board from a string, like B3, to a tuple, like (1, 2).