Struct chess::board::Board

pub struct Board {
    color: Color,
    pub board: Vec<Vec<Option<Piece>>>,
}

Represents the board state.

Fields

color

board

Methods

impl Board

fn new() -> Board

Initialize the board with starting positions.

fn get_piece(&self, cell: Cell) -> &Option<Piece>

Get the piece associated with a given cell index.

fn inbounds(&self, cell: &Cell) -> bool

fn is_empty(&self, cell: &Cell) -> bool

fn moves_until_collision(&self, dir: Cell, cell: Cell) -> Vec<Cell>

Returns the potential moves in a given direction, stopping upon a collision.

fn potential_moves(&self, cell: Cell) -> Vec<Cell>

Calculate the potential moves for a given cell index.

fn check(&self) -> bool

Helper function that checks for check

fn checkmate(&self) -> bool

Helper function that checks for checkmate

fn move_piece(&mut self, piece: Piece, to: Cell)

Helper function that moves a piece to a target cell

fn contains_enemy_king(&self, cells: Vec<Cell>) -> bool

Helper function that checks if a list of potential moves contains the king