Struct chess::Game

pub struct Game<'a> { /* fields omitted */ }

The game struct.

The coordinates used to access pieces are 0-indexed tuples of (usize, usize), and they follow the standard chess notation, so (0,0) corresponds to A1 in the bottom left corner, and (7,7) corresponds to H8 in the top right corner, seen from the white side.

The pieces are stored as Option<&Piece>, and are references to the pieces in the WHITE and BLACK array.

Eksamples

// Create a new game, with all pieces in their initial position.
let mut game = Game::new();

// The piece at C1 is supposed to be a bishop.
let bishop = game.get_from_pos((2,0));
if let Some(piece) = bishop {
    assert_eq!(piece.kind, Kind::Bishop);
    assert_eq!(piece.color, Color::White);
} else {
    panic!("The piece at A1 should be a bishop.");
}

Methods

impl<'a> Game<'a>

fn new() -> Game<'a>

Creates a new game, with all the pieces in the correct starting position.

Eksamples

let mut game = Game::new();

fn new_empty() -> Game<'a>

Creates a new game with an empty board.

Eksamples

let mut game = Game::new_empty();
assert_eq!(game.by_color(Color::White).len(), 0);
assert_eq!(game.by_color(Color::Black).len(), 0);

fn clear(&mut self)

Clears the board.

Eksamples

let mut game = Game::new();
assert_eq!(game.by_color(Color::White).len(), 16);

game.clear();
assert_eq!(game.by_color(Color::White).len(), 0);

fn ignore_kings(&mut self, ignore: bool)

Tells the game whether to ignore a lack of kings.

The game still sees if a possible move puts a king in check, but it no longer panics if one or both kings are missing. This can be useful when setting up special challenges.

Eksamples

let mut game = Game::new_empty();
game.set_at_pos((3, 3), Some(&WHITE[3]));

// This would cause a panic
// game.valid_moves((3, 3));

game.ignore_kings(true);
game.valid_moves((3, 3));

fn ignore_check(&mut self, ignore: bool)

Tells the game whether to ignore check tests.

Eksamples

let mut game = Game::new_empty();
// Put a black queen on D4 and a white king on C2
game.set_at_pos((3, 3), Some(&BLACK[4]));
game.set_at_pos((2, 1), Some(&WHITE[5]));

// With check tests in place, the king can only move to B1, B3 and C1.
assert_eq!(game.valid_moves((2, 1)).len(), 3);

// Whithout check tests the king can also move to B2, C3, D1, D2 and D3.
game.ignore_check(true);
assert_eq!(game.valid_moves((2, 1)).len(), 8);

fn get_from_pos(&self, pos: (usize, usize)) -> Option<&'a Piece>

Gets the piece at the given position on the board.

Returns an Option where Some contains a reference to the piece, and None means there was no piece at the given position.

Eksamples

let mut game = Game::new();

// Get the piece from A1
match game.get_from_pos((0, 0)) {
    Some(piece) => assert_eq!(piece.kind, Kind::Rook),
    None => panic!("There should be a rook here."),
}

// Returns None if the position is empty.
assert_eq!(game.get_from_pos((3, 4)), None);

fn set_at_pos(&mut self, pos: (usize, usize), piece: Option<&'a Piece>)

Sets the piece at the given position on the board.

The piece is passed as an Option, where the Some should contain a reference to the WHITE or BLACK arrays. Pass None to remove an existing piece.

Eksamples

let mut game = Game::new();

// Replace all white pawns with queens.
for i in 0..8 {
    game.set_at_pos((i, 1), Some(&WHITE[4]));
    match game.get_from_pos((i, 1)) {
        Some(piece) => {
            assert_eq!(piece.kind, Kind::Queen);
            assert_eq!(piece.color, Color::White);
        },
        None => panic!("There should be a queen here."),
    }
}

fn get_turn(&self) -> u32

Returns the current turn.

fn next_turn(&mut self)

Advances the game to the next turn.

fn by_color(&self, color: Color) -> Vec<((usize, usize), &'a Piece)>

Returns a vector of all pieces of a given color, and their position on the board.

The pieces are arrenged in the order they are found, starting at A1 through H1, then A2 through H2, until it reaches H8.

Eksamples

let game = Game::new();

// At the start of a chess game there should be 16 pieces of each color.
assert_eq!(game.by_color(Color::White).len(), 16);
assert_eq!(game.by_color(Color::Black).len(), 16);

// The 9th white piece should be the pawn at A2.
let pieces = game.by_color(Color::White);
assert_eq!(pieces[8].0, (0, 1));
assert_eq!(pieces[8].1.kind, Kind::Pawn);
assert_eq!(pieces[8].1.color, Color::White);

fn by_kind(&self, kind: Kind) -> Vec<((usize, usize), &'a Piece)>

Returns a vector of all pieces of a given kind, and their position on the board.

The pieces are arrenged in the order they are found, starting at A1 through H1, then A2 through H2, until it reaches H8.

Eksamples

let game = Game::new();

// At the start of a chess game there should be 16 pawns and 2 kings.
assert_eq!(game.by_kind(Kind::Pawn).len(), 16);
assert_eq!(game.by_kind(Kind::King).len(), 2);

// The 13th pawn should be the black pawn at E7..
let pawns = game.by_kind(Kind::Pawn);
assert_eq!(pawns[12].0, (4, 6));
assert_eq!(pawns[12].1.kind, Kind::Pawn);
assert_eq!(pawns[12].1.color, Color::Black);

fn by_kind_and_color(&self,
kind: Kind,
color: Color)
-> Vec<((usize, usize), &'a Piece)>

Returns a vector of all pieces of a given kind and color, and their position on the board.

The pieces are arrenged in the order they are found, starting at A1 through H1, then A2 through H2, until it reaches H8.

Eksamples

let game = Game::new();

// At the start of a chess game there should be 8 black pawns and 2 white knights.
assert_eq!(game.by_kind_and_color(Kind::Pawn, Color::Black).len(), 8);
assert_eq!(game.by_kind_and_color(Kind::Knight, Color::White).len(), 2);

// The 2nd black bishop should be at F8.
let bishops = game.by_kind_and_color(Kind::Bishop, Color::Black);
assert_eq!(bishops[1].0, (5, 7));
assert_eq!(bishops[1].1.kind, Kind::Bishop);
assert_eq!(bishops[1].1.color, Color::Black);

fn move_piece(&mut self,
from: (usize, usize),
to: (usize, usize))
-> Option<&'a Piece>

Moves a piece from one position to another.

The return value is an Option containing a reference to the captured piece (if any), or None if either of the positions given were empty. Trying to move from a position that doesn't contain a piece therefore returns None.

This function doesn't check whether the move is valid, only that the positions are within bounds. Therefore this should always be used together with valid_moves when playing proper chess.

Eksamples

let mut game = Game::new();

// Move a pawn from D2 to D3.
// This returns None because no pieces were captured.
assert_eq!(game.move_piece((3, 1), (3, 2)), None);

// The original position is now empty.
assert_eq!(game.get_from_pos((3, 1)), None);
// And the new position contains the pawn.
match game.get_from_pos((3, 2)) {
    Some(piece) => {
        assert_eq!(piece.kind, Kind::Pawn);
        assert_eq!(piece.color, Color::White);
    },
    None => panic!("There should be a pawn here."),
}

// Moving a pawn from D3 to H8 is illegal in chess, but can be done here.
// The captured rook is removed from the board, and returned.
let captured = game.move_piece((3, 2), (7, 7));
match captured {
    Some(piece) => {
        assert_eq!(piece.kind, Kind::Rook);
        assert_eq!(piece.color, Color::Black);
    },
    None => panic!("There should be a captured piece here."),
}

// There is no piece at B4, so trying to move from there is just returning None.
assert_eq!(game.move_piece((1, 3), (4, 0)), None);

fn move_pieces(&mut self,
moves: &[((usize, usize), (usize, usize))])
-> Option<&'a Piece>

Executes several moves, as stated in the given array.

The return value is Some containing the last captured piece (if any), or None if no pieces were captured or no pieces were moved. If one of the moves is out of bounds no moves are executed, and None is returned.

In cases where only one piece must be moved manually, move_piece is preferred.

This function is supposed to be called with the result of valid_moves. It is used instead of move_piece in case complex moves where several pieces is moved, like castling, is nessessary. This function doesn't check whether the moves are legal.

Eksamples

let mut game = Game::new();
let mut moves: Vec<((usize, usize), (usize, usize))>;

// Move a pawn from E2 forwards twice.
moves = vec![((4, 1), (4, 2)), ((4, 2), (4, 3))];
assert_eq!(game.move_pieces(&moves), None);
match game.get_from_pos((4, 3)) {
    Some(piece) => {
        assert_eq!(piece.kind, Kind::Pawn);
        assert_eq!(piece.color, Color::White);
    },
    None => panic!("There should be a pawn here."),
}

// When two pieces are captured only the last one is returned.
// Moves the pawn from E4, captures the queen at D8, then captures the rook at H8.
moves = vec![((4, 3), (3, 7)), ((3, 7), (7, 7))];
let captured = game.move_pieces(&moves);
match captured {
    Some(piece) => {
        assert_eq!(piece.kind, Kind::Rook);
        assert_eq!(piece.color, Color::Black);
    },
    None => panic!("There should be a rook here."),
}

fn valid_moves(&self,
pos: (usize, usize))
-> Vec<Vec<((usize, usize), (usize, usize))>>

Returns a vector of all the moves the piece at the given position can make.

The returned vector contains vectors of moves, as a tuple of the current location and the destination. This is done so that more complex moves that involve moving several pieces, such as castling, can be carried out. Each of these vectors can be passed to move_pieces to be executed.

If the given position doesn't contain a piece, a vector with size 0 is returned.

Eksamples

let mut game = Game::new();

// The pawn at E2 can only move forwards one or two squares.
let moves = game.valid_moves((4, 1));
// The returned vector contains two possible moves, each requiering only one move
// to be carried out.
assert_eq!(moves.len(), 2);
assert_eq!(moves[0].len(), 1);
assert_eq!(moves[1].len(), 1);

// The pawn can be moved to squares forwards.
assert_eq!(moves[0][0].0, (4, 1));
assert_eq!(moves[0][0].1, (4, 3));

// Or one step forwards.
assert_eq!(moves[1][0].0, (4, 1));
assert_eq!(moves[1][0].1, (4, 2));

// Lets move it two steps forwards, to E4.
game.move_pieces(&moves[0]);
assert!(game.get_from_pos((4, 3)) != None);
// Advance the turn. This is nessessary for some internal handling.
game.next_turn();

// Now we move a black pawn from D7 to D5.
for v in game.valid_moves((3, 6)) {
    if v.len() == 1 && v[0].1 == (3, 4) {
        game.move_pieces(&v);
        assert!(game.get_from_pos((3, 4)) != None);
        game.next_turn();
        break;
    }
}

// Now the white pawn can capture the black pawn at D5.
for v in game.valid_moves((4, 3)) {
    if v.len() == 1 && v[0].1 == (3, 4) {
        assert!(game.move_pieces(&v) != None);
        match game.get_from_pos((3, 4)) {
            Some(piece) => {
                assert_eq!(piece.kind, Kind::Pawn);
                assert_eq!(piece.color, Color::White);
            },
            None => panic!("There should be a piece here."),
        }
        game.next_turn();
        break;
    }
}

fn in_check(&self, color: Color) -> bool

Sees whether the king of the given color is currently in check or not.

Eksamples

// Clear the board, then put a black king at C5, and a white pawn at D4.
let mut game = Game::new_empty();
game.set_at_pos((3, 3), Some(&WHITE[0]));
game.set_at_pos((2, 4), Some(&BLACK[5]));

assert!(game.in_check(Color::Black));

fn check_victory(&self) -> Option<(Victory, Color)>

Checks whether the game is won, and returns the victory type and the color of the victor, or None if the game isn't won yet. In case of a draw a random color is returned.

Eksamples

// Clear the board, then put a black king at A1, and a white queen at B2. We also need a
// white king on the board, if not the program panics.
let mut game = Game::new_empty();
game.set_at_pos((1, 1), Some(&WHITE[4]));
game.set_at_pos((0, 0), Some(&BLACK[5]));
game.set_at_pos((6, 7), Some(&WHITE[5]));

// The king is in check, but it can still move and take out the queen.
assert_eq!(game.check_victory(), None);

// Move the queen to B3
game.set_at_pos((1, 2), Some(&WHITE[4]));
game.set_at_pos((1, 1), None);

// Now the king isn't in check, but the king can't move so it's a stalemate draw.
assert_eq!(game.check_victory(), Some((Victory::Stalemate, Color::White)));

// Add another queen at C3
game.set_at_pos((2, 2), Some(&WHITE[4]));

// Now the king is in check, and can't move, so white has won by checkmate.
assert_eq!(game.check_victory(), Some((Victory::Checkmate, Color::White)));

fn move_to_an(&self,
m: &[((usize, usize), (usize, usize))],
result: bool,
unicode: bool)
-> String

Turns a move, as returned from valid_moves, into algebraic notation (AN).

If result is true the function will detect whether a checkmate or a stalemate has occured, and add "1-0", "0-1" or "½-½".

If unicode is true the pieces are represented by unicode symbols instead of letters. Only black pieces are used, as they are easier to see.

Eksamples

let mut game = Game::new();

// Turn a string from AN into a move, and back into AN.
let mut m = game.an_to_move("Nc3", Color::White).unwrap();
assert_eq!(game.move_to_an(&m, true, false), "Nc3");
assert_eq!(game.move_to_an(&m, true, true), "♞c3");

m = game.an_to_move("e5", Color::Black).unwrap();
assert_eq!(game.move_to_an(&m, true, false), "e5");

fn an_to_move(&self,
s: &str,
color: Color)
-> Option<Vec<((usize, usize), (usize, usize))>>

Turns a string in algebraic notation (AN) into a move that can be passed to move_pieces.

This function supports abbreviated algebraic notation, which means that certain characters can be removed, as long as it is unambiguous. For example, an 'x' (which signals a capture) is completely ignored, and can even be added to moves that doesn't end with a capture. The same goes for '=Q' (which signals a pawn promotion) and 'e.p.' (which signals en passant).

The pieces can be represented by both letters and unicode symbols.

To get the proper algebraic notation instead of the abbreviated one from a user, pass the result of an_to_move to move_to_an.

This function returns None both if the input is malformed and if the move is invalid. There is currently no way to distinguish the two.

Eksamples

let mut game = Game::new();
// Start by moving a knight from B1 to C3.
let mut m = game.an_to_move("Nc3", Color::White);
assert_eq!(m, Some(vec![((1, 0), (2, 2))]));
game.move_pieces(&m.unwrap());

// Move a black pawn from D7 to D5.
m = game.an_to_move("d5", Color::Black);
assert_eq!(m, Some(vec![((3, 6), (3, 4))]));
game.move_pieces(&m.unwrap());

m = game.an_to_move("e4", Color::White);
assert_eq!(m, Some(vec![((4, 1), (4, 3))]));
game.move_pieces(&m.unwrap());

// Now the pawn at D5 can capture the pawn at E4.
m = game.an_to_move("dxe4", Color::Black);
assert_eq!(m, Some(vec![((3, 4), (4, 3))]));
// Abbreviated notation is also valid.
assert_eq!(m, game.an_to_move("de4", Color::Black));
assert_eq!(m, game.an_to_move("de", Color::Black));
game.move_pieces(&m.unwrap());

// Fast-forwards a little.
m = game.an_to_move("Nf3", Color::White);
game.move_pieces(&m.unwrap());
m = game.an_to_move("Ng5", Color::White);
game.move_pieces(&m.unwrap());

// Now both white knights can reach E4, so "Ne4" isn't enough.
m = game.an_to_move("Ne4", Color::White);
assert_eq!(m, None);

// ...so we must specify the file the knight is moving from.
m = game.an_to_move("Nce4", Color::White);
assert_eq!(m, Some(vec![((2, 2), (4, 3))]));

// We could also specify the rank, or both the rank and the file.
assert_eq!(m, game.an_to_move("N3e4", Color::White));
assert_eq!(m, game.an_to_move("Nc3e4", Color::White));
game.move_pieces(&m.unwrap());

// Fast forwards some more.
m = game.an_to_move("Qf3", Color::White);
game.move_pieces(&m.unwrap());
m = game.an_to_move("Be2", Color::White);
game.move_pieces(&m.unwrap());
m = game.an_to_move("b3", Color::White);
game.move_pieces(&m.unwrap());
m = game.an_to_move("Bb2", Color::White);
game.move_pieces(&m.unwrap());

// Kingside castling.
m = game.an_to_move("0-0", Color::White);
assert_eq!(m, Some(vec![((4, 0), (5, 0)), ((5, 0), (6, 0)), ((7, 0), (5, 0))]));

// Queenside castling.
m = game.an_to_move("0-0-0", Color::White);
assert_eq!(m, Some(vec![((4, 0), (3, 0)), ((3, 0), (2, 0)), ((0, 0), (3, 0))]));

fn move_to_string(&self, m: &((usize, usize), (usize, usize))) -> String

Turns a move tuple into a human readable description.

Eksamples

let game = Game::new();
let m = game.valid_moves((4, 1));
assert_eq!(game.move_to_string(&m[0][0]), "Moving white pawn from E2 to E4");

fn moves_to_string(&self, m: &[((usize, usize), (usize, usize))]) -> String

Turns an array of move tuples, like entries returned from valid_moves, into a human readable description.

Eksamples

let game = Game::new();
let m = game.valid_moves((4, 1));
assert_eq!(game.moves_to_string(&m[0]), "Moving white pawn from E2 to E4");

fn board_to_string(&self, unicode: bool) -> String

Returns the game board as a string.

Set unicode to true if you want the pieces represented by their unicode symbols instead of letters. If unicode is false the same letters that are used in algebraic notation is used, with the addition of 'P' for pawns. The white pieces are represented by uppercase letters, while black are lowercase.

Eksamples

let game = Game::new();
let mut board = game.board_to_string(false);
assert_eq!(board,
           "rnbqkbnr\
          \npppppppp\
          \n        \
          \n        \
          \n        \
          \n        \
          \nPPPPPPPP\
          \nRNBQKBNR");

board = game.board_to_string(true);
assert_eq!(board,
           "♜♞♝♛♚♝♞♜\
          \n♟♟♟♟♟♟♟♟\
          \n        \
          \n        \
          \n        \
          \n        \
          \n♙♙♙♙♙♙♙♙\
          \n♖♘♗♕♔♗♘♖");

fn three_fold_repetition(&self) -> bool

Checks whether there has occured a three fold repetition.

fn fifty_move_rule(&self) -> bool

Checks whether a player can invoke the fifty-move-rule